Use a string as a dep in useEffect to avoid having it changed always
See https://blog.logrocket.com/frustrations-with-react-hooks/
React uses Object.is to compare if a dependency changed compared to the
previous values. And the `keys` dependency was created like
`Objects.values(...)`, so that a new object was passed in all the time
and `Object.is()` always saw it as "changed". So the key handler always
got re-registered and sometimes got "lost"...