Add connect and discover actions
This allows reacting on hub events that are fired immediately after a
hub got connected (like "attach" events), but before the CONNECT action took
place. When a hub is discovered, it's immediately added to the "hubs",
but with a state "connected=false".
The CONNECT action then sets the hub to "connected=true".

The tilt indicators are only shown, when the TILT port got attached.